Add usernames on their products

+++ index.php
@@ -3,6 +3,7 @@ session_start();
 require 'autoload.php';
 ?>
 
+
 <?php if (!isset($_SESSION['username'])) : ?>
   <a href="register.php">Register</a>
   <a href="login.php">Login</a>
@@ -23,11 +24,14 @@ $products = $pdo->view();
 foreach ($products as $product) :
 ?>
   <img src="<?php echo $product['image'] ?>" alt="Image" height="50px">
+  <b><?php echo @$_SESSION['username'] == $product['username'] ? 'You' : $product['username'] ?></b>
   <h1><?php echo $product['title'] ?></h1>
   <p><?php echo $product['description'] ?></p>
 
-  <button><a href="update_product.php?id=<?php echo $product['id'] ?>">Update</a></button>
-  <button><a href="includes/delete_product.php?id=<?php echo $product['id'] ?>">Delete</a></button>
+  <?php if (@$_SESSION['username'] == $product['username']) : ?>
+    <button><a href="update_product.php?id=<?php echo $product['id'] ?>">Update</a></button>
+    <button><a href="includes/delete_product.php?id=<?php echo $product['id'] ?>">Delete</a></button>
+  <?php endif ?>
   <br>
   <br>
 <?php endforeach ?>
